Permissions part II

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken')
 const { randomBytes } = require('crypto')
 const { promisify } = require('util')
 const { transport, makeANiceEmail } = require('../mail')
+const { hasPermission } = require('../utils')
 
 const mutations = {
 	async createItem(parent, args, ctx, info) {
@@ -45,7 +46,15 @@ const mutations = {
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id }
 
-		const item = await ctx.db.query.item({ where }, `{ id title }`)
+		const item = await ctx.db.query.item({ where }, `{ id title user }`)
+
+		// Check if the user can delete the item
+		const ownsItem = item.user.id === ctx.request.userId
+		const hasPermissions = ctx.request.user.permissions.some(permission =>
+			['ADMIN', 'ITEMDELETE'].includes(permission)
+		)
+
+		if (!ownsItem && !hasPermission) throw new Error('Your are not allowed!')
 
 		return ctx.db.mutation.deleteItem({ where }, info)
 	},
@@ -175,6 +184,34 @@ const mutations = {
 		})
 
 		return updatedUser
+	},
+	async updatePermissions(parent, args, ctx, info) {
+		if (!ctx.request.userId) throw new Error('You must be logged in!')
+
+		const currentUser = await ctx.db.query.user(
+			{
+				where: {
+					id: ctx.request.userId
+				}
+			},
+			info
+		)
+
+		hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+
+		return ctx.db.mutation.updateUser(
+			{
+				data: {
+					permissions: {
+						set: args.permissions // because is an array
+					}
+				},
+				where: {
+					id: args.userId //it could be someone else id
+				}
+			},
+			info
+		)
 	}
 }
